test(runtime): move surviving coverage to resident owners
