+++ src/runtime/src/runtime/transaction/tests/program/mod.rs
@@ -1,9 +1,3 @@
 mod effects;
-mod explicit;
-mod extension_failures;
-mod implicit;
-mod integrity;
-mod poisoning;
-mod rollback;
 mod savepoints;
 mod support;
